(chore) Removed console.logs and moved chart.jsx, a redux component into containers folder
