+++ src/components/data-circles.jsx
@@ -27,7 +27,6 @@ const renderCircles = (props) => {
 };
 
 export default (props) => {
-  console.log(props);
   return (
     <g>{ props.data.map(renderCircles(props)) }
