+++ src/components/posts_index.jsx
@@ -1,10 +1,8 @@
 import React, { Component } from 'react';
-import Chart from './chart';
+import Chart from '../containers/chart';
 
 export default class Test extends Component {
   render() {
-    console.log(this.context);
-    console.log(Chart);
     return (
       <div>
         <h1> Hello, world! </h1>
